feat(settings): add infrastructure layer and refactor repository
- Created infrastructure layer for settings feature.
- Renamed LlmSettingsRepository to SettingsRepository for consistency.
- Moved repository implementation and data source to infrastructure/.
- Updated use cases, application layer, and tests to reflect architectural changes.
- Ensured proper DI wiring with SettingsRepositoryImpl.
