+++ lib/features/settings/domain/usecases/export_data.dart
@@ -1,7 +1,7 @@
-import '../repositories/llm_settings_repository.dart';
+import '../repositories/settings_repository.dart';
 
 class ExportData {
-  final LlmSettingsRepository repository;
+  final SettingsRepository repository;
   ExportData(this.repository);
 
   Future<String> call() => repository.exportData();
