+++ lib/features/settings/domain/usecases/get_app_settings.dart
@@ -1,8 +1,8 @@
 import '../entities/app_settings.dart';
-import '../repositories/llm_settings_repository.dart';
+import '../repositories/settings_repository.dart';
 
 class GetAppSettings {
-  final LlmSettingsRepository repository;
+  final SettingsRepository repository;
   GetAppSettings(this.repository);
 
   Future<AppSettings?> call() => repository.getAppSettings();
