+++ lib/features/settings/domain/usecases/import_data.dart
@@ -1,7 +1,7 @@
-import '../repositories/llm_settings_repository.dart';
+import '../repositories/settings_repository.dart';
 
 class ImportData {
-  final LlmSettingsRepository repository;
+  final SettingsRepository repository;
   ImportData(this.repository);
 
   Future<void> call(String data) => repository.importData(data);
