+++ lib/features/settings/infrastructure/models/llm_config_dto.dart
