+++ test/features/settings/infrastructure/datasources/settings_local_datasource_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:isar/isar.dart';
-import 'package:orionhealth_health/features/settings/data/datasources/settings_local_datasource.dart';
+import 'package:orionhealth_health/features/settings/infrastructure/datasources/settings_local_datasource.dart';
 import 'package:orionhealth_health/features/settings/domain/entities/app_settings.dart';
 import 'package:orionhealth_health/features/settings/domain/entities/llm_config.dart';
