+++ test/features/settings/infrastructure/repositories/settings_repository_impl_test.dart
@@ -3,12 +3,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:isar/isar.dart';
 import 'package:orionhealth_health/features/settings/domain/entities/app_settings.dart';
 import 'package:orionhealth_health/features/settings/domain/entities/llm_config.dart';
-import 'package:orionhealth_health/features/settings/data/repositories/llm_settings_repository_impl.dart';
-import 'package:orionhealth_health/features/settings/data/datasources/settings_local_datasource.dart';
+import 'package:orionhealth_health/features/settings/data/repositories/settings_repository_impl.dart';
+import 'package:orionhealth_health/features/settings/infrastructure/datasources/settings_local_datasource.dart';
 
 void main() {
   late Isar isar;
-  late LlmSettingsRepositoryImpl repository;
+  late SettingsRepositoryImpl repository;
   late Directory tempDir;
 
   setUpAll(() async {
@@ -27,14 +27,14 @@ void main() {
       [LlmConfigSchema, AppSettingsSchema],
       directory: tempDir.path,
     );
-    repository = LlmSettingsRepositoryImpl(SettingsLocalDataSource(isar));
+    repository = SettingsRepositoryImpl(SettingsLocalDataSource(isar));
   });
 
   tearDown(() async {
     await isar.close(deleteFromDisk: true);
   });
 
-  group('LlmSettingsRepositoryImpl', () {
+  group('SettingsRepositoryImpl', () {
     test('getLlmConfig returns null when no config exists', () async {
       final result = await repository.getLlmConfig();
       expect(result, isNull);
